Feat : update UserService -> mypage, deleteUser

+++ src/main/java/com/weave/weaveserver/dto/UserResponse.java
@@ -19,7 +19,7 @@ public class UserResponse {
         private String email;
         private String image;
         private String loginType;
-        private int countTeam;
+        private Long countTeam;
     }
 
 }
